[wasm] Debugger tests - always use the same sdk on helix, as the one
.. used for building.

This failed when `global.json` specified a preview4 sdk, which was used
to build the tests. But the built tests (including the preview4 testhost) were
send to helix with the preview2 (version set by arcade -
`DotNetCliVersion`), which causes:

```
Testhost process exited with error: It was not possible to find any compatible framework version
The framework 'Microsoft.NETCore.App', version '7.0.0-preview.4.22201.3' (x64) was not found.
  - The following frameworks were found:
      7.0.0-preview.2.22152.2 at [C:\helix\work\correlation\dotnet-cli\shared\Microsoft.NETCore.App]
You can resolve the problem by installing the specified framework and/or SDK.
```