Make ReRegisterForFinalize GC test more robust (#67896)
There is no guarantee that the JIT does not introduce a temporary for
the read of the fields `obj1` and `obj2` when they are passed to the
function calls, which might keep them alive even when nulling out the
fields. This happens when CFG is enabled, causing the test to fail.

Change the test to access the fields in separate methods marked as
noinline to fix the problem.