[mono] Remove "old" sgen bridge (#67957)
We default to the tarjan bridge for quite a while now, so remove the "old" bridge. We decided to keep the "new" bridge since there are still a few usecases for it (some Xamarin.Android customers use it). If a user selects the "old" bridge we'll use the "new" bridge instead and print a warning.