Improving STJ source generator support for record types  (#68064)
* Improving STJ source generator support for record types

Improving the analyzer to not flag a warning about init-only
properties that are actually constructor parameters. In particular,
record types with positional parameters desugar to classes with both
a constructor and init-only properties (eg, to support "with"
expressions). Since these properties are deserialized via the
constructor and not the property setter, they support the fast path and
round trip just fine, so the diagonstic analyzer should not raise the
SYSLIB1037 warning.

Issue 58770

* Improving STJ source generator support for record types

Edit to fix missing semi colon in unit test's CreateCompiliation string

Issue 58770

* Case insensitive matching of props with ctor params

* Additional unit tests

* Test for source gen positional records roundtrip

Co-authored-by: Joe Harjung <jharjung@integratedlabor.com>