Only build the GC simulator tests in the GC simulator leg (#68177)
* Only build the GC simulator tests in the GC simulator leg

Only build the GC simluator tests instead of all Pri1 tests. When we build all Pri1 tests in this leg, nearly all tests are skipped as only the GCSimluator tests are actually run. These skipped tests leave a ton of files on disk, causing issues in Helix. By only building the actual tests we are going to run, we can avoid this additional infrastructure cost.

* Harden the conditional statements against empty merged wrapper set

* Filter tests during "copy native binaries" step as we also generate the run scripts in this step.

* Fix indentation

* Move filter arg up

* No dash for tree arg in windows

* Specially handle sending one assembly with PayloadGroup0

Co-authored-by: Tomas <trylek@microsoft.com>