JIT: fix bad arm64 move peephole (#68631)
`IF_DR_2G` `(mov sp/reg, reg/sp)` has an implicit reg2, so an unwary peephole opt
might mistakenly think it is `mov reg, x0`. If the opt is unlucky and the next
instruction is a real `mov reg, x0` the peephole opt might suppress the second
instruction, leaving the wrong value in `reg`.

Fix is to just disallow `IF_DR_2G` in the peephole.

Fixes #68527.