Mitigate stress hangs by marking GC-sensitive tests as out-of-proc (#68733)
According to Andrew Au's findings in the issue

https://github.com/dotnet/runtime/issues/68060

a few tests explicitly manipulate the GC finalizer queue in a way
that is prone to hangs due to entering an infinite finalization
loop under GC stress. While the primary issue tracks the problem
in its entirety, the fix is non-trivial as it's not yet fully
understood whether we should fix the runtime, the GC stress mode
behavior or the tests themselves; after all, quite a few other
tests use explicit finalization too. To mitigate the problem in
the short term and help cleaning up the pending failures in GC
stress runs I propose marking tests that are known to
regularly cause these issues as requiring process isolation
to run out-of-process as they used to before the test merging.

Thanks

Tomas