Do not fold mismatched `OBJ(ADDR(LCL_VAR))` (#69271)
* Do not fold mismatched OBJ(ADDR(LCL_VAR))

The transforms in question could fold, e. g. "OBJ<8>(ADDR(LCL_VAR<16>))"
to just the "LCL_VAR", necessitating workarounds in block morphing that
had to rematerialize the block node to restore IR's validity.

It is better to simply not create questionable IR in the first place.

* Add a test

* Re-enable tests

* Fix regressions

By using layout compatibility in "gtNewStructVal" instead of handle equality.

This is safe to do because "gtNewStructVal" is only used in contexts of block
copies (as opposed to call arguments).