Fix W^X executable memory releasing on Unix (#69697)
The PAL_VirtualReserveFromExecutableMemoryAllocatorWithinRange API that
the ExecutableAllocator uses on Unix to reserve VA space for executable
memory was registering the allocated memory range in the PAL in the list
of allocated memory blocks. But the ExecutableAllocator is releasing
that memory just using unmap, so the block remained registered in that
list after the unmap. That resulted in an assert in the sanity check of
the list when the PAL virtual memory allocator got the VA range that was
unmapped, but found it in the list.

The fix is to add a parameter to the
PAL_VirtualReserveFromExecutableMemoryAllocatorWithinRange to specify
whether the allocated memory should be added to the list of allocation
info or not.