Optimize better with maximum array length (#69735)
* Optimize better with maximum array length

The maximum array length is 0x7FFFFFC7 (see
https://github.com/dotnet/runtime/pull/43301), which is slightly less than
the largest 32-bit signed integer. Use this value in range check and assertion
prop to optimize better. This is due do knowing that a value smaller than
the maximum array length plus a small constant will not overflow. This leads
to being able to remove some bounds checks.

* Use CORINFO_Array_MaxLength

* Consider both Span and array for bounds check limit

* Formatting