Remove Or, And, Not + related fixes (#69839)
* Remove unordered Or, And and Not.

The SymbolicRegexNodeKind elements Or, And and Not as well as any related classes/functions have been removed.
This includes SymbolicRegexSet, TransitionRegex and SymbolicNFA.

SaveDGML, which depended on Or, has been reworked by pulling out the exploration logic into
SymbolicRegexMatcher.Explore, which does the exploration in the main transition structures of the matcher.

GenerateRandomMembers is now SampleMatches and has also been rewritten to use the transition logic from
SymbolicRegexMatcher.

SymbolicRegexMatcher is now a partial class and SaveDGML, Explore, and SampleMatches are implemented in
separate files as additions to the matcher.

* Remove negative sampling

Negative sampling in GenerateRandomMembers depended on Not.

* Rename OrderedOr to Alternate

* Fix and generalize FixedLengthMarker support

Logic to extract the fixed length out of an end state is now in SymbolicRegexNode.ResolveFixedLength. The new function
supports finding the fixed length marker on the path that the backtracking matcher would accept a match and also
supports conditional nullability.

Logic to add fixed length markers in RegexNodeConverter is now in a function
SymbolicRegexNode.AddFixedLengthMarkers, which made it easier to make more general.